requirements.txt
opencv-python
face-recognition
numpy
mysql-connector-python
twilio
streamlit